refactor(codes): use shared code value for claims
Why:
- AssuranceClaimCode became a thin wrapper after code value validation moved into the shared contract.
- The code catalog contract should have one canonical value object for the shared code shape.

What Changed:
- Convert UcliCodeValue from a validator-only helper into the shared code value object.
- Use UcliCodeValue directly for ready and compile claim codes.
- Remove the redundant AssuranceClaimCode type and its dedicated tests.
- Move UcliCodeValue under the code catalog contract area and centralize its invalid-value message.

Impact:
- Claim code JSON remains unchanged.
- Code value validation and value-object behavior now have a single implementation.
