+++ src/Ucli.Application/Features/Assurance/Ready/ReadyClaimCodes.cs
@@ -3,15 +3,15 @@ namespace MackySoft.Ucli.Application.Features.Assurance.Ready;
 /// <summary> Defines assurance claim codes emitted by the <c>ready</c> command. </summary>
 internal static class ReadyClaimCodes
 {
-    public static readonly AssuranceClaimCode UnityReadyExecution = new("UNITY_READY_EXECUTION");
+    public static readonly UcliCodeValue UnityReadyExecution = new("UNITY_READY_EXECUTION");
 
-    public static readonly AssuranceClaimCode UnityReadyMutation = new("UNITY_READY_MUTATION");
+    public static readonly UcliCodeValue UnityReadyMutation = new("UNITY_READY_MUTATION");
 
-    public static readonly AssuranceClaimCode UnityReadyTest = new("UNITY_READY_TEST");
+    public static readonly UcliCodeValue UnityReadyTest = new("UNITY_READY_TEST");
 
-    public static readonly AssuranceClaimCode UnityReadyReadIndex = new("UNITY_READY_READ_INDEX");
+    public static readonly UcliCodeValue UnityReadyReadIndex = new("UNITY_READY_READ_INDEX");
 
-    public static IReadOnlyList<AssuranceClaimCode> All { get; } =
+    public static IReadOnlyList<UcliCodeValue> All { get; } =
     [
         UnityReadyExecution,
         UnityReadyMutation,
@@ -20,7 +20,7 @@ internal static class ReadyClaimCodes
     ];
 
     /// <summary> Gets the claim code for one ready target. </summary>
-    public static AssuranceClaimCode ForTarget (ReadyTarget target)
+    public static UcliCodeValue ForTarget (ReadyTarget target)
     {
         return target switch
         {
